Improvements

+++ src/tuning/convertBinToEpd.nim
@@ -14,7 +14,7 @@ proc binToEpd*(inFileName, outFileName: string, maxLen = int.high) =
   while not inFileStream.atEnd:
     let
       position = inFileStream.readPosition
-      value = inFileStream.readFloat64
+      value = inFileStream.readFloat32
 
     if position[king, white].countSetBits != 1 or position[king, black].countSetBits != 1:
       continue
